move InlineStyles in utils
